requirements.txt
# PDF parsing
pdfplumber>=0.10.0
# Or use pypdf>=3.0.0 as backup

# AWS S3 (Step 3 use)
boto3>=1.28.0

# Pinecone (Step 3 use)
pinecone-client>=2.2.0

# Embedding (Gemini)
google-generativeai>=0.6.0

# Other tools
pyyaml>=6.0  # To read documents_config.yaml
python-dotenv>=1.0.0  # To load .env file
